LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/data-uri-expected.txt

CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:text/plain,Hello, World!. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:text/plain;base64,SGVsbG8sIFdvcmxkIQ==. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:text/html,Hello, World!. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:text/html;charset=UTF-8,Hello, World!. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:image/png,Hello, World!. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:text/plain,Hello, World!. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load data:text/plain,Hello, World!. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: XMLHttpRequest cannot load data:text/plain,Hello, World!. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load data:text/plain,Hello, World!. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load data:text/plain,Hello, World!. Preflight response is not successful

FAIL XHR method GET with charset text/plain assert_equals: expected "Hello, World!" but got ""
FAIL XHR method GET with charset text/plain (base64) assert_equals: expected "Hello, World!" but got ""
FAIL XHR method GET with charset text/html assert_equals: expected "Hello, World!" but got ""
FAIL XHR method GET with charset text/html;charset=UTF-8 assert_equals: expected "Hello, World!" but got ""
FAIL XHR method GET with charset image/png assert_equals: expected "Hello, World!" but got ""
PASS XHR method POST with charset text/plain 
PASS XHR method PUT with charset text/plain 
PASS XHR method DELETE with charset text/plain 
PASS XHR method HEAD with charset text/plain 
PASS XHR method UNICORN with charset text/plain 
